correccion decimales

+++ SisteamasVentas/SistemasVentas.Modelos/DetalleVenta.cs
@@ -8,8 +8,8 @@ public class DetalleVenta
 	public int IdVenta { get; set; }
 	public int IdProducto { get; set; }
 	public int Cantidad { get; set; }
-	public int PrecioVenta { get; set; }
-	public int SubTotal { get; set; }
+	public decimal PrecioVenta { get; set; }
+	public decimal SubTotal { get; set; }
 	public string Estado { get; set; }
 	}
 }
